RecommendationRequestService interface was added.
